+++ monarch_utils/lib/log_config.dart
@@ -1,2 +1,1 @@
 export 'src/log_config/log_config.dart';
-export 'src/log_config/console_log_config.dart';